bookplace-backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary
drf-spectacular==0.27.2
